Separate VerifiedContractsBuffer from VerifiedContractsByAccountIdCache.
